Convert targets to use buildable folder references
This is the new default behaviour in Xcode 16. From the Xcode release notes:

Minimize project file changes, and avoid version control conflicts with buildable folder references.
Convert an existing group to a buildable folder with the “Convert to Folder” context menu item in the Project Navigator. Buildable folders only record the folder path into the project file without enumerating the contained files. This minimizes diffs to the project when files are added and removed, and avoids source control conflicts with your team.

https://developer.apple.com/documentation/xcode-release-notes/xcode-16-release-notes